refactor: simulation관련 로직을 훅으로 분리
