+++ frontend/src/pages/ContestPage.tsx
@@ -1,16 +1,15 @@
 import { css } from '@style/css';
 
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 import ContestBreadCrumb from '@/components/Contest/ContestBreadCrumb';
 import Editor from '@/components/Editor/Editor';
 import ProblemViewer from '@/components/Problem/ProblemViewer';
 import SimulatorList from '@/components/Simulation/SimulatorList';
-import type { TestCase } from '@/components/Simulation/types';
 import SubmissionResult from '@/components/SubmissionResult';
 import { SITE } from '@/constants';
+import { useSimulations } from '@/hooks/simulation/useSimulations';
 import mockData from '@/mockData.json';
-import evaluator from '@/modules/evaluator';
 
 const notFoundProblem = {
   title: 'Problem Not Found',
@@ -26,64 +25,22 @@ const INITIAL_PROBLEM_ID = 6;
 
 export default function ContestPage() {
   const CONTEST_NAME = 'Test'; // api로 받을 정보
-
+  const { testCases, runSimulation, changeParam } = useSimulations();
   const [currentProblemId] = useState(INITIAL_PROBLEM_ID);
   const targetProblem =
     mockData.problems.find((problem) => problem.id === currentProblemId) || notFoundProblem;
   const [code, setCode] = useState<string>(targetProblem.solutionCode);
-  const [testCases, setTestCases] = useState<TestCase[]>([
-    { param: '', result: '' },
-    { param: '', result: '' },
-    { param: '', result: '' },
-    { param: '', result: '' },
-    { param: '', result: '' },
-  ]);
 
   const handleChangeCode = (newCode: string) => {
     setCode(newCode);
   };
 
-  useEffect(() => {
-    return evaluator.subscribe(({ result, error, task }) => {
-      if (!task) return;
-
-      const taskId = task.clientId;
-
-      setTestCases((oldTestCases) => {
-        return oldTestCases.map((tc, index) => {
-          if (index !== taskId) return tc;
-
-          if (error) {
-            return {
-              ...tc,
-              result: `${error.name}: ${error.message} \n${error.stack}`,
-            };
-          }
-          return {
-            ...tc,
-            result,
-          };
-        });
-      });
-    });
-  }, []);
-
   const handleTestExec = () => {
-    setTestCases((oldTestCases) => {
-      return oldTestCases.map(toEvaluatingState);
-    });
-
-    const tasks = testCases.map((tc, index) => evaluator.createEvalMessage(index, code, tc.param));
-
-    evaluator.evaluate(tasks);
+    runSimulation(code);
   };
 
   const handleChangeParam = (index: number, newParam: string) => {
-    const changedTestCase = testCases.find((_, i) => i === index);
-    if (changedTestCase) {
-      changedTestCase.param = newParam;
-    }
-    setTestCases([...testCases]);
+    changeParam(index, newParam);
   };
 
   const crumbs = [SITE.NAME, CONTEST_NAME, targetProblem.title];
@@ -111,13 +68,6 @@ export default function ContestPage() {
     </main>
   );
 }
-
-const toEvaluatingState = (testcase: TestCase) => {
-  return {
-    ...testcase,
-    result: '계산중...',
-  };
-};
 
 const style = css({
   backgroundColor: '#1e1e1e',
